+++ frontend/src/hooks/index.ts
@@ -2,7 +2,13 @@ export { queryKeys } from './queryKeys'
 export { useCardExpandAnimation } from './useCardExpandAnimation'
 export { CartProvider, useCart } from './useCart'
 export { useCompleteMission, useMissions } from './useMissions'
-export { useSpacecraft, useSpacecrafts, useUpdateCargo, useUpdateSpacecraftStatus } from './useSpacecrafts'
+export {
+  useSpacecraft,
+  useSpacecrafts,
+  useUpdateCargo,
+  useUpdateFuel,
+  useUpdateSpacecraftStatus
+} from './useSpacecrafts'
 export { useStation, useTransferCargo } from './useStation'
 export { usePurchase, useStoreProducts } from './useStore'
 export {